cast enum to float to fix (really bad) warning
git-svn-id: http://skia.googlecode.com/svn/trunk@9316 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ bench/BitmapBench.cpp
@@ -171,7 +171,7 @@ protected:
         return 1;
 #else
         return (paint.getFlags() & SkPaint::kBicubicFilterBitmap_Flag) ?
-                BICUBIC_DUR_SCALE : 1;
+                (float)BICUBIC_DUR_SCALE : 1;
 #endif
     }
